+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.6.4 (WIP)
 
+* Fix the server-side `Request` class implementation
+
 ## 0.6.3 (2018-12-06)
 
 * Add `RURL#usedBaseProperty(name)` to find out if a property came from the base location