+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add `cast` argument to `Array#unique()`
 * `Pledge#_started` is now a timestamp, not a boolean
 * Add duration properties to the `Pledge` class
+* Add `Blast.REPLACE_OPEN_TAG_NEWLINE` symbol for use in `String#stripTags()`
 
 ## 0.6.3 (2018-12-06)